+++ src/tasks/tui/edit.py
@@ -2,8 +2,8 @@
 
 from textual.app import ComposeResult
 from textual.screen import ModalScreen
-from textual.widgets import Header, Footer, Static, Input, Label
-from textual.containers import Grid
+from textual.widgets import Header, Footer, Input, Label, Checkbox
+from textual.containers import Vertical, Horizontal
 
 from tasks.models import Task
 
@@ -23,17 +23,36 @@ class EditScreen(ModalScreen[Task]):
         self.__task = task
 
     def compose(self) -> ComposeResult:
-        yield Grid(
-            Header(),
-            Label("Name:"),
-            Input(value=self.__task.name, id="name"),
+        with Vertical(id="dialog"):
+            yield Header()
+            with Horizontal():
+                yield Label("Done:")
+                yield Checkbox(value=self.__task.done, id="done")
+            with Horizontal():
+                yield Label("Name:")
+                yield Input(value=self.__task.name, id="name")
+            with Horizontal():
+                yield Label("Priority:")
+                yield Input(value=str(self.__task.priority), id="priority", type="integer")
+            with Horizontal():
+                yield Label("Description:")
+                yield Input(value=self.__task.description, id="description")
+
             # TODO: Add more content 
-            Footer(), id="dialog")
+            yield Footer()
     
     def _update_task(self) -> None:
-        name_input: Input = self.query_one("#name")
+        done_checkbox = self.query_one("#done", Checkbox)
+        name_input = self.query_one("#name", Input)
+        priority_input = self.query_one("#priority", Input)
+        description_input = self.query_one("#description", Input)
 
         self.__task.name = name_input.value
+        self.__task.done = done_checkbox.value
+        self.__task.priority = int(priority_input.value)
+        self.__task.description = description_input.value
+
+
 
     def action_escape(self) -> None:
         self._update_task()
